requirements.txt

google-cloud
google-api-python-client==1.6.2 
pytest
